Moved app.py to FoodWasteApp/ and removed duplicate version
